requirement.txt
numpy-stl
vtk
patchelf
termcolor
torchgeometry
scikit-image
numpy
scipy
ipdb
joblib
opencv-python==4.6.0.66
tqdm 
pyyaml
wandb
mujoco>=2.3.7
scikit-image
gym
git+https://github.com/ZhengyiLuo/smplx.git@master
lxml
autograd
scikit-learn
chumpy
patchelf
wandb
pyvirtualdisplay
patchelf
lxml
ipdb
chardet
cchardet
geoopt
imageio-ffmpeg
easydict
open3d
hydra-core 
gymnasium
dm_env